PassManager: Refactor / separate into a module
This PR moves code in preparation for a full fledged pass manager for the IR.

Reference: https://github.com/FuelLabs/sway/discussions/3596
