+++ sway-ir/src/bin/opt.rs
@@ -4,7 +4,7 @@ use std::{
 };
 
 use anyhow::anyhow;
-use sway_ir::{error::IrError, function::Function, optimize, Context};
+use sway_ir::{ConstCombinePass, DCEPass, InlinePass, Mem2RegPass, PassManager, SimplifyCfgPass};
 
 // -------------------------------------------------------------------------------------------------
 
@@ -64,159 +64,6 @@ fn write_to_output<S: Into<String>>(ir_str: S, path_str: &Option<String>) -> std
                 .write(ir_str.into().as_ref())
                 .map(|_| ())
         }
-    }
-}
-
-// -------------------------------------------------------------------------------------------------
-
-trait NamedPass {
-    fn name() -> &'static str;
-    fn descr() -> &'static str;
-    fn run(ir: &mut Context) -> Result<bool, IrError>;
-
-    fn run_on_all_fns<F: FnMut(&mut Context, &Function) -> Result<bool, IrError>>(
-        ir: &mut Context,
-        mut run_on_fn: F,
-    ) -> Result<bool, IrError> {
-        let funcs = ir
-            .module_iter()
-            .flat_map(|module| module.function_iter(ir))
-            .collect::<Vec<_>>();
-        let mut modified = false;
-        for func in funcs {
-            if run_on_fn(ir, &func)? {
-                modified = true;
-            }
-        }
-        Ok(modified)
-    }
-}
-
-type NamePassPair = (&'static str, fn(&mut Context) -> Result<bool, IrError>);
-
-#[derive(Default)]
-struct PassManager {
-    passes: HashMap<&'static str, NamePassPair>,
-}
-
-impl PassManager {
-    fn register<T: NamedPass>(&mut self) {
-        self.passes.insert(T::name(), (T::descr(), T::run));
-    }
-
-    fn run(&self, name: &str, ir: &mut Context) -> Result<bool, IrError> {
-        self.passes.get(name).expect("Unknown pass name!").1(ir)
-    }
-
-    fn contains(&self, name: &str) -> bool {
-        self.passes.contains_key(name)
-    }
-
-    fn help_text(&self) -> String {
-        let summary = self
-            .passes
-            .iter()
-            .map(|(name, (descr, _))| format!("  {name:16} - {descr}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-
-        format!("Valid pass names are:\n\n{summary}",)
-    }
-}
-
-// -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-
-struct InlinePass;
-
-impl NamedPass for InlinePass {
-    fn name() -> &'static str {
-        "inline"
-    }
-
-    fn descr() -> &'static str {
-        "inline function calls."
-    }
-
-    fn run(ir: &mut Context) -> Result<bool, IrError> {
-        // For now we inline everything into `main()`.  Eventually we can be more selective.
-        let main_fn = ir
-            .module_iter()
-            .flat_map(|module| module.function_iter(ir))
-            .find(|f| f.get_name(ir) == "main")
-            .unwrap();
-        optimize::inline_all_function_calls(ir, &main_fn)
-    }
-}
-
-// -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-
-struct ConstCombinePass;
-
-impl NamedPass for ConstCombinePass {
-    fn name() -> &'static str {
-        "constcombine"
-    }
-
-    fn descr() -> &'static str {
-        "constant folding."
-    }
-
-    fn run(ir: &mut Context) -> Result<bool, IrError> {
-        Self::run_on_all_fns(ir, optimize::combine_constants)
-    }
-}
-
-// -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-
-struct SimplifyCfgPass;
-
-impl NamedPass for SimplifyCfgPass {
-    fn name() -> &'static str {
-        "simplifycfg"
-    }
-
-    fn descr() -> &'static str {
-        "merge or remove redundant blocks."
-    }
-
-    fn run(ir: &mut Context) -> Result<bool, IrError> {
-        Self::run_on_all_fns(ir, optimize::simplify_cfg)
-    }
-}
-
-// -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-
-struct DCEPass;
-
-impl NamedPass for DCEPass {
-    fn name() -> &'static str {
-        "dce"
-    }
-
-    fn descr() -> &'static str {
-        "Dead code elimination."
-    }
-
-    fn run(ir: &mut Context) -> Result<bool, IrError> {
-        Self::run_on_all_fns(ir, optimize::dce)
-    }
-}
-
-// -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-
-struct Mem2RegPass;
-
-impl NamedPass for Mem2RegPass {
-    fn name() -> &'static str {
-        "mem2reg"
-    }
-
-    fn descr() -> &'static str {
-        "Promote local memory to SSA registers."
-    }
-
-    fn run(ir: &mut Context) -> Result<bool, IrError> {
-        Self::run_on_all_fns(ir, optimize::promote_to_registers)
     }
 }
