+++ sway-ir/src/optimize/dce.rs
@@ -5,9 +5,25 @@
 //!   2. At the time of inspecting a definition, if it has no uses, it is removed.
 //! This pass does not do CFG transformations. That is handled by simplify_cfg.
 
-use crate::{Block, Context, Function, Instruction, IrError, Module, Value, ValueDatum};
+use crate::{Block, Context, Function, Instruction, IrError, Module, NamedPass, Value, ValueDatum};
 
 use std::collections::{HashMap, HashSet};
+
+pub struct DCEPass;
+
+impl NamedPass for DCEPass {
+    fn name() -> &'static str {
+        "dce"
+    }
+
+    fn descr() -> &'static str {
+        "Dead code elimination."
+    }
+
+    fn run(ir: &mut Context) -> Result<bool, IrError> {
+        Self::run_on_all_fns(ir, dce)
+    }
+}
 
 fn can_eliminate_instruction(context: &Context, val: Value) -> bool {
     let inst = val.get_instruction(context).unwrap();
